Remove obsolete iotivity 0.9.1 bbappend
The latest meta-oic that brings us IoTivity 0.9.2 had removed the
recipe for iotivity 0.9.1. That breaks the build since bitbake is
unable to find a match for iotivity_0.9.1.bbappend.

Additionally, IoTivity 0.9.2/meta-oic no longer depends on cereal
and has improved (but not fixed!) the dependency for downloading
external sources. Therefore, it's safe to remove the iotivity
0.9.1 bbappend.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>