Change OPENJDK_OE_CONF variable to include optimizations
Currently OPENJDK_OE_CONF-variable, which is used in do_compile()-step, does
not include optimization flags. This raises a warning during build as
FORTIFY_SOURCES feature requires optimizations to work. This warning is then
interpreted as error due to -Werror.

This commit only changes the --with-extra-cflags field in OPENJDK_OE_CONF string
to include ${SELECTED_OPTIMIZATION}-variable to provide the desired optimization
level. Other flags are not changed.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>