libgcrypt: fix the patch headers to avoid combo-layer breakage
The combo-layer tool we use to pull in patches from individual
layers uses '---' mark to separate patches and the commit
messages.

The libgcrypt patches in turn have some strange '--' in the commit
headers which become '---' when git revert && git format-patch
is run. This results in a race where combo-layer puts it's
added meta information in a wrong place and the patch no longer
applies.

Fix by removing the strange '--' chars from the patches.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>