u-boot: Fix beaglebone bootloader
U-boot bootloader was broken by an upstream patch and subsequently fixed in
another patch. The oe-core u-boot upgrade from 2015.10 to 2016.01 includes the
former but not the latter, so the fix patch is backported.

This patch has been submitted to oe-core, but fixing the build for M3 would be
nice so that we do not provide broken images.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>