meta-java: Do not use absolute paths when bootstrapping
Both cacao-initial-native and ecj-initial-native use hardcoded absolute
paths to various files. Normally this is not an issue, but it makes it
difficult to share sstate between otherwise identical workers, if the
file paths are not identical.

This is fixed by forming the paths at runtime, which allows using the
files in different workers with different folders.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>