recipes-core/glibc: lifted over glibc customization from meta-appfw.
Lifted over a small glibc customization bbappend from meta-appfw.
It simply splits out getent to a subpackage of its own. getent is
a hard requirement for systemd-nspawn and we split getent out of
its previous subpackage so that we can pull it in without the rest
of the binaries in the same subpackage.

Everything is just an exact verbatim copy of what we've already
been doing, we just lifted it over, as agreed, to a common place
for easier tracking.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>